Changed the site a little to separate the page that loads depending on is a user is admin or not. Still need to make UserHome and RepairAdmin display different info. UserHome should only display devices from that user, with no admin functions, API also protected from non-admin calls.
RepairAdmin is for techs. Will add more bools to SiteUser later to split
admins and techs. Goal is for customer to have their portal, still need to tackle non-user (look up by email or force them to register, or if they have account,
 but were not logged in then have another app to trasfer repairs under their email to their username). The tech/admin side sees everything and can manage everthing. Also, want an undo option that will undo last DB action.

Also, added /admin route and the withAuth HOC to protech it. Want to make another HOC called withAdmin that checks token and is user is admin.